requirements.txt
notion-client==0.9.0
python-binance==1.0.15